feat(sidebar): pull out search component
re #384
